requirements.txt
aiofiles==0.8.0
albumentations==1.2.1
anyio==3.6.1
argcomplete==2.0.0
backports.cached-property==1.0.2
boto3==1.24.42
botocore==1.27.42
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cycler==0.11.0
dacite==1.6.0
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
eventlet==0.33.1
fiftyone==0.16.5
fiftyone-brain==0.8.2
fiftyone-db==0.3.0
fonttools==4.34.4
future==0.18.2
glob2==0.7
graphql-core==3.1.7
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
imageio==2.20.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.4.4
labelImg==1.8.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mongoengine==0.20.0
motor==2.5.1
ndjson==0.3.1
networkx==2.8.5
numpy==1.23.1
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
patool==1.12
Pillow==9.2.0
plotly==4.14.3
pprintpp==0.4.0
priority==2.0.0
psutil==5.9.1
pycocotools==2.0.4
Pygments==2.12.0
pymongo==3.12.3
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
retrying==1.3.3
rfc3986==1.5.0
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.0
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sse-starlette==0.10.3
sseclient-py==1.7.2
starlette==0.16.0
strawberry-graphql==0.96.0
tabulate==0.8.10
threadpoolctl==3.1.0
tifffile==2022.7.28
toml==0.10.2
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.11
voxel51-eta==0.7.1
wrapt==1.14.1
wsproto==1.1.0
xmltodict==0.13.0
